refactor(settings): colocate section components inside their tab folders
