+++ src/phone/apps/settings/component/tab-app/accessibility-section.vue
