+++ src/phone/apps/settings/component/tab-app/index.vue
@@ -1,8 +1,8 @@
 <script setup lang="ts">
 import { inject } from 'vue'
 import SectionList from '@/components/layout-kit/section-list.vue'
-import AudioSection from '../app-settings/audio-section.vue'
-import AccessibilitySection from '../app-settings/accessibility-section.vue'
+import AudioSection from './audio-section.vue'
+import AccessibilitySection from './accessibility-section.vue'
 import SettingsBackButton from '../settings-back-button.vue'
 import SettingsSaveButton from '../settings-save-button.vue'
 import { settingsLayoutKey } from '../../layout'
