+++ src/phone/apps/settings/component/tab-subscription/index.vue
@@ -3,8 +3,8 @@ import { useI18n } from 'vue-i18n'
 import UiButton from '@/components/ui-kit/button.vue'
 import SectionList from '@/components/layout-kit/section-list.vue'
 import LabeledSection from '@/components/layout-kit/labeled-section.vue'
-import PlanSection from '../billing-settings/plan-section.vue'
-import PaymentMethodsSection from '../billing-settings/payment-methods-section.vue'
+import PlanSection from './plan-section.vue'
+import PaymentMethodsSection from './payment-methods-section.vue'
 import SettingsBackButton from '../settings-back-button.vue'
 import { useMemberStore } from '@/stores/member'
 import { useSubscriptionQuery } from '@/api/billing'
